+++ lilith/src/main/java/de/huxhorn/lilith/swing/table/LoggingEventViewTable.java
@@ -1,6 +1,6 @@
 /*
  * Lilith - a log event viewer.
- * Copyright (C) 2007-2011 Joern Huxhorn
+ * Copyright (C) 2007-2017 Joern Huxhorn
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,6 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+
 package de.huxhorn.lilith.swing.table;
 
 import de.huxhorn.lilith.data.logging.LoggingEvent;
@@ -51,18 +52,20 @@ import javax.swing.table.TableColumn;
 public class LoggingEventViewTable
 	extends EventWrapperViewTable<LoggingEvent>
 {
-	public static final String DEFAULT_COLUMN_NAME_ID = "ID";
-	public static final String DEFAULT_COLUMN_NAME_TIMESTAMP = "Timestamp";
-	public static final String DEFAULT_COLUMN_NAME_LEVEL = "Level";
-	public static final String DEFAULT_COLUMN_NAME_LOGGER_NAME = "Logger";
-	public static final String DEFAULT_COLUMN_NAME_MESSAGE = "Message";
-	public static final String DEFAULT_COLUMN_NAME_THROWABLE = "Throwable";
-	public static final String DEFAULT_COLUMN_NAME_THREAD = "Thread";
-	public static final String DEFAULT_COLUMN_NAME_MARKER = "Marker";
-	public static final String DEFAULT_COLUMN_NAME_NDC = "NDC";
-	public static final String DEFAULT_COLUMN_NAME_APPLICATIION = "Application";
-	public static final String DEFAULT_COLUMN_NAME_CONTEXT = "Context";
-	public static final String DEFAULT_COLUMN_NAME_SOURCE = "Source";
+	private static final long serialVersionUID = 2969867582530249593L;
+
+	private static final String DEFAULT_COLUMN_NAME_ID = "ID";
+	private static final String DEFAULT_COLUMN_NAME_TIMESTAMP = "Timestamp";
+	private static final String DEFAULT_COLUMN_NAME_LEVEL = "Level";
+	private static final String DEFAULT_COLUMN_NAME_LOGGER_NAME = "Logger";
+	private static final String DEFAULT_COLUMN_NAME_MESSAGE = "Message";
+	private static final String DEFAULT_COLUMN_NAME_THROWABLE = "Throwable";
+	private static final String DEFAULT_COLUMN_NAME_THREAD = "Thread";
+	private static final String DEFAULT_COLUMN_NAME_MARKER = "Marker";
+	private static final String DEFAULT_COLUMN_NAME_NDC = "NDC";
+	private static final String DEFAULT_COLUMN_NAME_APPLICATION = "Application";
+	private static final String DEFAULT_COLUMN_NAME_CONTEXT = "Context";
+	private static final String DEFAULT_COLUMN_NAME_SOURCE = "Source";
 
 	public LoggingEventViewTable(MainFrame mainFrame, EventWrapperTableModel<LoggingEvent> model, boolean global)
 	{
@@ -86,7 +89,7 @@ public class LoggingEventViewTable
 			new ThrowableTooltipGenerator());
 		tooltipGenerators.put(DEFAULT_COLUMN_NAME_TIMESTAMP,
 			new TimestampTooltipGenerator());
-		tooltipGenerators.put(DEFAULT_COLUMN_NAME_APPLICATIION,
+		tooltipGenerators.put(DEFAULT_COLUMN_NAME_APPLICATION,
 			new ApplicationTooltipGenerator());
 		tooltipGenerators.put(DEFAULT_COLUMN_NAME_CONTEXT,
 			new ContextTooltipGenerator());
@@ -153,7 +156,7 @@ public class LoggingEventViewTable
 		}
 		{
 			TableColumn col = new TableColumn(0);
-			col.setHeaderValue(DEFAULT_COLUMN_NAME_APPLICATIION);
+			col.setHeaderValue(DEFAULT_COLUMN_NAME_APPLICATION);
 			col.setCellRenderer(new ApplicationRenderer());
 			tableColumns.put(col.getHeaderValue(), col);
 		}
@@ -185,7 +188,7 @@ public class LoggingEventViewTable
 		result.add(new PersistentTableColumnModel.TableColumnLayoutInfo(DEFAULT_COLUMN_NAME_THREAD, 75, true));
 		result.add(new PersistentTableColumnModel.TableColumnLayoutInfo(DEFAULT_COLUMN_NAME_MARKER, 75, true));
 		result.add(new PersistentTableColumnModel.TableColumnLayoutInfo(DEFAULT_COLUMN_NAME_NDC, 75, true));
-		result.add(new PersistentTableColumnModel.TableColumnLayoutInfo(DEFAULT_COLUMN_NAME_APPLICATIION, 75, true));
+		result.add(new PersistentTableColumnModel.TableColumnLayoutInfo(DEFAULT_COLUMN_NAME_APPLICATION, 75, true));
 		result.add(new PersistentTableColumnModel.TableColumnLayoutInfo(DEFAULT_COLUMN_NAME_CONTEXT, 75, true));
 		result.add(new PersistentTableColumnModel.TableColumnLayoutInfo(DEFAULT_COLUMN_NAME_SOURCE, 75, isGlobal()));
 
@@ -194,8 +197,8 @@ public class LoggingEventViewTable
 
 	public void saveLayout()
 	{
-		List<PersistentTableColumnModel.TableColumnLayoutInfo> infos = tableColumnModel.getColumnLayoutInfos();
-		mainFrame.getApplicationPreferences().writeLoggingColumnLayout(isGlobal(), infos);
+		List<PersistentTableColumnModel.TableColumnLayoutInfo> layoutInfoList = tableColumnModel.getColumnLayoutInfos();
+		mainFrame.getApplicationPreferences().writeLoggingColumnLayout(isGlobal(), layoutInfoList);
 	}
 
 	protected List<PersistentTableColumnModel.TableColumnLayoutInfo> loadLayout()